requirements.txt
web3
kafka-python
pyahocorasick
ujson
clickhouse-driver